+++ docs/changelog.rst
@@ -15,6 +15,7 @@ New features
 ------------
 
 * :class:`aspecd.tasks.FigurereportTask` for creating figure captions that can, *e.g.*, be included in other documents
+* Attributes ``labelspacing`` and ``fontsize`` in :class:`aspecd.plotting.LegendProperties`
 
 
 Version 0.6.4